requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.14
langgraph==0.2.16
langchain-openai==0.1.22
langchain-google-genai==1.0.10
mcp-use==0.1.5
pydantic==2.7.4
python-dotenv==1.0.1
httpx==0.27.0